+++ v16/messages/get_configuration.go
@@ -13,6 +13,6 @@ type GetConfigurationResponse struct {
 
 type ConfigurationKey struct {
 	Key      string  `json:"key" validate:"required,max=50"`
-	Readonly bool    `json:"readonly" validate:"required"`
+	Readonly bool    `json:"readonly"`
 	Value    *string `json:"value,omitempty" validate:"omitempty,max=500"`
 }